set dtype for out of bounds slice

+++ napari/layers/image/_slice.py
@@ -211,7 +211,10 @@ class _ImageSliceRequest:
     def __call__(self) -> _ImageSliceResponse:
         if self._slice_out_of_bounds():
             return _ImageSliceResponse.make_empty(
-                slice_input=self.slice_input, rgb=self.rgb, request_id=self.id
+                slice_input=self.slice_input,
+                rgb=self.rgb,
+                request_id=self.id,
+                dtype=self.data.dtype,
             )
         with self.dask_indexer():
             return (
